Change consult_price from FloatField to DecimalField.

+++ users/models.py
@@ -31,7 +31,10 @@ class Doctor(models.Model):
                                   max_length=3,
                                   verbose_name='科室')
     years = models.PositiveIntegerField(verbose_name='从医时间')
-    consult_price = models.FloatField(default=0, verbose_name='咨询价格')
+    consult_price = models.DecimalField(default=0,
+                                        max_digits=10,
+                                        decimal_places=2,
+                                        verbose_name='咨询价格')
     title = models.CharField(choices=TITLE_CHOICES, max_length=1, verbose_name='职称')
     ratings = models.PositiveIntegerField(default=5, verbose_name='评分')
     patient_num = models.PositiveIntegerField(default=0, verbose_name='已帮助患者')
